# openapi.yaml
openapi: 3.1.0
info:
  description: OAuth 2.0 and OpenID Connect identity for Minecraft Java Edition accounts.
  license:
    name: MIT
  title: CraftLogin API
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      bearerFormat: opaque
      scheme: bearer
      type: http
    oauth2:
      flows:
        authorizationCode:
          authorizationUrl: https://craftlogin.com/oauth2/authorize
          scopes:
            offline_access: Maintain delegated access with refresh-token rotation.
            openid: Authenticate the Minecraft account.
            profile: Read the current Minecraft username and avatar.
          tokenUrl: https://craftlogin.com/oauth2/token
      type: oauth2
  schemas:
    craftlogin.error-response:
      additionalProperties: false
      properties:
        error:
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
          required:
            - code
            - message
          type: object
      required:
        - error
      type: object
...
paths:
  /api/avatars/{identifier}/skin:
    get:
      summary: Get a Minecraft skin
      tags:
        - Avatars
      description: Return the current signed skin for a Minecraft player or texture hash.
      parameters:
        - schema:
            pattern: ^(?:[A-Za-z0-9_]{3,16}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-fA-F]{32}|[0-9a-fA-F]{64})$
            type: string
          in: path
          name: identifier
          required: true
      responses:
        "200":
          description: Default Response
          headers:
            cache-control:
              schema:
                type: string
            etag:
              schema:
                type: string
          content:
            image/png:
              schema:
                format: binary
                type: string
        "304":
          description: Default Response
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        default:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
  /api/avatars/{identifier}/capes:
    get:
      summary: List available Minecraft capes
      tags:
        - Avatars
      description: List available capes across supported providers (Mojang, OptiFine,
        LabyMod, MinecraftCapes, 5zig, SkinMC).
      parameters:
        - schema:
            pattern: ^(?:[A-Za-z0-9_]{3,16}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-fA-F]{32}|[0-9a-fA-F]{64})$
            type: string
          in: path
          name: identifier
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.capes-response"
        "304":
          description: Default Response
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        default:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
  /api/avatars/{identifier}/face:
    get:
      summary: Render a Minecraft face
      tags:
        - Avatars
      description: Render the front eight-by-eight face with its transparent Minecraft
        head overlay.
      parameters:
        - schema:
            default: all
            enum:
              - base
              - all
            type: string
          in: query
          name: layers
          required: false
        - schema:
            enum:
              - classic
              - slim
            type: string
          in: query
          name: model
          required: false
          description: Arm model for texture-hash subjects (player subjects always use
            their signed profile model).
        - schema:
            enum:
              - 5zig
              - any
              - labymod
              - minecraftcapes
              - mojang
              - optifine
              - skinmc
            type: string
          in: query
          name: provider
          required: false
          description: Cape provider for cape-bearing views (back, duo, wings); defaults
            to any.
        - schema:
            default: "128"
            enum:
              - "32"
              - "64"
              - "128"
              - "256"
            type: string
          in: query
          name: size
          required: false
        - schema:
            pattern: ^(?:[A-Za-z0-9_]{3,16}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-fA-F]{32}|[0-9a-fA-F]{64})$
            type: string
          in: path
          name: identifier
          required: true
      responses:
        "200":
          description: Default Response
          headers:
            cache-control:
              schema:
                type: string
            etag:
              schema:
                type: string
          content:
            image/png:
              schema:
                format: binary
                type: string
        "304":
          description: Default Response
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        default:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
  /api/avatars/{identifier}/bust:
    get:
      summary: Render a Minecraft bust
      tags:
        - Avatars
      description: Render a flat front view of the head, torso, and arms.
      parameters:
        - schema:
            default: all
            enum:
              - base
              - all
            type: string
          in: query
          name: layers
          required: false
        - schema:
            enum:
              - classic
              - slim
            type: string
          in: query
          name: model
          required: false
          description: Arm model for texture-hash subjects (player subjects always use
            their signed profile model).
        - schema:
            enum:
              - 5zig
              - any
              - labymod
              - minecraftcapes
              - mojang
              - optifine
              - skinmc
            type: string
          in: query
          name: provider
          required: false
          description: Cape provider for cape-bearing views (back, duo, wings); defaults
            to any.
        - schema:
            default: "128"
            enum:
              - "32"
              - "64"
              - "128"
              - "256"
            type: string
          in: query
          name: size
          required: false
        - schema:
            pattern: ^(?:[A-Za-z0-9_]{3,16}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-fA-F]{32}|[0-9a-fA-F]{64})$
            type: string
          in: path
          name: identifier
          required: true
      responses:
        "200":
          description: Default Response
          headers:
            cache-control:
              schema:
                type: string
            etag:
              schema:
                type: string
          content:
            image/png:
              schema:
                format: binary
                type: string
        "304":
          description: Default Response
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        default:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
  /api/avatars/{identifier}/body:
    get:
      summary: Render a Minecraft body
      tags:
        - Avatars
      description: Render a flat front view of the full player from its current signed skin.
      parameters:
        - schema:
            default: all
            enum:
              - base
              - all
            type: string
          in: query
          name: layers
          required: false
        - schema:
            enum:
              - classic
              - slim
            type: string
          in: query
          name: model
          required: false
          description: Arm model for texture-hash subjects (player subjects always use
            their signed profile model).
        - schema:
            enum:
              - 5zig
              - any
              - labymod
              - minecraftcapes
              - mojang
              - optifine
              - skinmc
            type: string
          in: query
          name: provider
          required: false
          description: Cape provider for cape-bearing views (back, duo, wings); defaults
            to any.
        - schema:
            default: "128"
            enum:
              - "32"
              - "64"
              - "128"
              - "256"
            type: string
          in: query
          name: size
          required: false
        - schema:
            pattern: ^(?:[A-Za-z0-9_]{3,16}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-fA-F]{32}|[0-9a-fA-F]{64})$
            type: string
          in: path
          name: identifier
          required: true
      responses:
        "200":
          description: Default Response
          headers:
            cache-control:
              schema:
                type: string
            etag:
              schema:
                type: string
          content:
            image/png:
              schema:
                format: binary
                type: string
        "304":
          description: Default Response
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        default:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
  /api/avatars/{identifier}/back:
    get:
      summary: Render a Minecraft back view
      tags:
        - Avatars
      description: Render a flat back view of the full player, including the worn cape
        when the account has one.
      parameters:
        - schema:
            default: all
            enum:
              - base
              - all
            type: string
          in: query
          name: layers
          required: false
        - schema:
            enum:
              - classic
              - slim
            type: string
          in: query
          name: model
          required: false
          description: Arm model for texture-hash subjects (player subjects always use
            their signed profile model).
        - schema:
            enum:
              - 5zig
              - any
              - labymod
              - minecraftcapes
              - mojang
              - optifine
              - skinmc
            type: string
          in: query
          name: provider
          required: false
          description: Cape provider for cape-bearing views (back, duo, wings); defaults
            to any.
        - schema:
            default: "128"
            enum:
              - "32"
              - "64"
              - "128"
              - "256"
            type: string
          in: query
          name: size
          required: false
        - schema:
            pattern: ^(?:[A-Za-z0-9_]{3,16}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-fA-F]{32}|[0-9a-fA-F]{64})$
            type: string
          in: path
          name: identifier
          required: true
      responses:
        "200":
          description: Default Response
          headers:
            cache-control:
              schema:
                type: string
            etag:
              schema:
                type: string
          content:
            image/png:
              schema:
                format: binary
                type: string
        "304":
          description: Default Response
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        default:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
  /api/avatars/{identifier}/side:
    get:
      summary: Render a Minecraft side profile
      tags:
        - Avatars
      description: Render a flat left-side profile of the full player.
      parameters:
        - schema:
            default: all
            enum:
              - base
              - all
            type: string
          in: query
          name: layers
          required: false
        - schema:
            enum:
              - classic
              - slim
            type: string
          in: query
          name: model
          required: false
          description: Arm model for texture-hash subjects (player subjects always use
            their signed profile model).
        - schema:
            enum:
              - 5zig
              - any
              - labymod
              - minecraftcapes
              - mojang
              - optifine
              - skinmc
            type: string
          in: query
          name: provider
          required: false
          description: Cape provider for cape-bearing views (back, duo, wings); defaults
            to any.
        - schema:
            default: "128"
            enum:
              - "32"
              - "64"
              - "128"
              - "256"
            type: string
          in: query
          name: size
          required: false
        - schema:
            pattern: ^(?:[A-Za-z0-9_]{3,16}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-fA-F]{32}|[0-9a-fA-F]{64})$
            type: string
          in: path
          name: identifier
          required: true
      responses:
        "200":
          description: Default Response
          headers:
            cache-control:
              schema:
                type: string
            etag:
              schema:
                type: string
          content:
            image/png:
              schema:
                format: binary
                type: string
        "304":
          description: Default Response
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        default:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
...
  /api/avatars/{identifier}/wings:
    get:
      summary: Render Minecraft elytra wings
      tags:
        - Avatars
      description: Render a flat back view of the full player with deployed elytra
        wings, drawn from the wing region of the worn cape texture.
      parameters:
        - schema:
            default: all
            enum:
              - base
              - all
            type: string
          in: query
          name: layers
          required: false
        - schema:
            enum:
              - classic
              - slim
            type: string
          in: query
          name: model
          required: false
          description: Arm model for texture-hash subjects (player subjects always use
            their signed profile model).
        - schema:
            enum:
              - 5zig
              - any
              - labymod
              - minecraftcapes
              - mojang
              - optifine
              - skinmc
            type: string
          in: query
          name: provider
          required: false
          description: Cape provider for cape-bearing views (back, duo, wings); defaults
            to any.
        - schema:
            default: "128"
            enum:
              - "32"
              - "64"
              - "128"
              - "256"
            type: string
          in: query
          name: size
          required: false
        - schema:
            pattern: ^(?:[A-Za-z0-9_]{3,16}|[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}|[0-9a-fA-F]{32}|[0-9a-fA-F]{64})$
            type: string
          in: path
          name: identifier
          required: true
      responses:
        "200":
          description: Default Response
          headers:
            cache-control:
              schema:
                type: string
            etag:
              schema:
                type: string
          content:
            image/png:
              schema:
                format: binary
                type: string
        "304":
          description: Default Response
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "429":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "500":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        "503":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
        default:
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/craftlogin.error-response"
...
servers:
  - url: https://craftlogin.com
...